+++ CHANGELOG.md
@@ -30,6 +30,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 * The [`$NO_COLOR` environment variable](https://no-color.org/) no longer
   overrides the `ui.color` configuration if explicitly set.
 
+* `jj edit` has been renamed to `jj touchup`.
+
 * `jj git push` no longer aborts if you attempt to push an open commit (but it
   now aborts if a commit does not have a description).